chore: split oversized file to comply with model size constraints
